pyproject.toml
[tool.poetry]
name = "egit"
version = "0.1.0"
description = ""
authors = ["Aissa Ouboukioud <aissa.1bacsm1@gmail.com>"]
readme = "README.md"

[tool.poetry.scripts]
egit="egit.main:main"

[tool.poetry.dependencies]
python = "^3.12"
rich = "^13.9.4"
gitpython = "^3.1.43"
typer = "^0.13.0"


[tool.poetry.group.dev.dependencies]
ipython = "^8.29.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
